Remove SUSE Linux 9.3 compatibility to improve error messages (bsc#1172578)
For an example profile looking like this:

```xml
<profile>
  <general>...</general>
  <report>...</report>
  <users>...</users>
</profile
```

SUSE Linux 9.3 required the elements to be categorized as "install" or
"configure", like this:

```xml
<profile>
  <install>
    <general>...</general>
    <report>...</report>
  </install>
  <configure>
    <users>...</users>
  </configure>
</profile
```

but that complicates the schema just enough so that xmllint error messages for
an invalid profile are *completely* useless, instead of just confusing.

This simplification will fix the error message when an unknown child of the
root element is seen, and on lower leves will report at least the parent
element of the culprit (not ideal but helps).